requirements.txt
asgiref==3.10.0
blinker==1.9.0
branca==0.8.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Django==5.2.7
folium==0.20.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.4
overpy==0.7
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
Werkzeug==3.1.3
xyzservices==2025.10.0
